requirements.txt
attrs
mypy